create CardFoodImage component
