+++ app/(tabs)/scan/index.tsx
@@ -3,11 +3,10 @@ import { Stack } from 'expo-router';
 import HomeButton from '@/components/Buttons/Home';
 import CameraUploader from '@/components/Scan/CameraUploader';
 import { useState } from 'react';
-import { Button, Card, Text, Spinner } from 'tamagui';
-import { Image } from "expo-image";
 import axios from 'axios';
 import { EnvConfig } from '@/providers/EnvConfig';
 import { base64Image } from '@/helpers/diverse';
+import CardFoodImage from '@/components/Scan/CardFoodImage';
 
 export default function ScanIndex(){
 
@@ -41,77 +40,16 @@ export default function ScanIndex(){
       />
 
       <ScrollView style={{ flex: 1 }}>
-        <CameraUploader setUri={setUri} uri={uri} />
-
-        <Card elevate bordered margin="$2">
-          <View style={styles.cardRow}>
-            {/* left content */}
-            <View style={styles.imageContainer}>
-              <Image
-                source={uri ? {uri} : require('@/assets/images/no-image.png')}
-                contentFit="cover"
-                style={styles.imagePreview}
-              />
-            </View>
-
-            {/* right content */}
-            <View style={styles.rightContent}>
-              <View >
-                <Text  style={styles.textInstruction} >
-                  { uri ? 'The image was successfully uploaded.' : 'Please upload an image to proceed.'}
-                </Text>
-                <Text style={styles.textInstruction} >
-                  { uri ? 'Press the button if you want to receive the analysis.' : ''}
-                </Text>
-              </View>
-
-              {uri && <View style={styles.buttonContainer}>
-                <Button
-                  width={80}
-                  size="$3"
-                  themeInverse
-                  onPress={analyzeImage}
-                >
-                  Analyze
-                  {isLoading ? <Spinner color="grey" /> : null}
-                </Button>
-              </View>
-              }
-            </View>
-          </View>
-        </Card>
-
+        <CameraUploader
+          setUri={setUri}
+          uri={uri}
+        />
+        <CardFoodImage
+          uri={uri}
+          isLoading={isLoading}
+          analyzeImage={analyzeImage}
+        />
       </ScrollView>
     </>
   );
 }
-
-const styles = StyleSheet.create({
-  cardRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  imageContainer: {
-    flex: 1,
-    paddingRight: 10,
-  },
-  imagePreview: {
-    width: '100%',
-    height: 160,
-    borderRadius: 10,
-  },
-  rightContent: {
-    flex: 1,
-    paddingHorizontal: 5,
-    justifyContent: 'center',
-  },
-  textInstruction: {
-    fontSize: 13,
-    marginBottom: 5
-  },
-  buttonContainer: {
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-    marginTop: 10,
-  },
-})
